Updated NSTimeIntervalTypeTransform to be Int64
- Moved transformTest to appropriate file
